requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
oauth2client==4.0.0
packaging==16.8
passlib==1.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
PyJWT==1.4.2
pymongo==3.4.0
pyparsing==2.2.0
rsa==3.4.2
six==1.10.0
Werkzeug==0.11.15
